==> 1_gpt_df.ipynb <==
import pandas as pd

df = pd.read_csv('database/AI_social_media.csv') #https://huggingface.co/datasets/newsmediabias/Social_media_cleaned_data?row=1
df['origin'] = 'AI'
df['model'] = 'chat gpt'
df = df.drop(columns = ['label_bias'])
df.head()

df.isna().sum()
df.duplicated().sum()

df = df.dropna(subset=['text'])
df = df.drop_duplicates()

gpt = pd.read_csv("database/gpt2.csv")
gpt.head()

gpt.isna().sum()
gpt = gpt.dropna(subset=['#YIAYplan Use my awesome collection of Amiibos to destroy all in my path.'])

gpt.duplicated().sum()
gpt = gpt.drop_duplicates()

gpt['origin'] = 'AI'
gpt['model'] = 'chat gpt'
gpt = gpt.rename(columns = {'#YIAYplan Use my awesome collection of Amiibos to destroy all in my path.':'text'})
gpt = gpt.drop(columns = ['Unnamed: 1'])
gpt.head()

df = pd.concat([df, gpt], axis=0)
df.info()

df['origin'].value_counts()

db = [
    "climate_gpt3_10k.csv", "climate_gpt3_15k.csv", "covid_gpt3_10k.csv", "covid_gpt3_15k.csv",
    "politics_gpt3_10k.csv", "politics_gpt3_15k.csv", "science_gpt3_10k.csv",
]

# Carrega o primeiro dataset para inicializar `data`
data = pd.read_csv("database/science_gpt3_15k.csv")

# Itera sobre os demais arquivos (evita duplicar o já carregado)
for i in db:
    if i != "science_gpt3_15k.csv":
        gpt = pd.read_csv(f"database/{i}")
        data = pd.concat([data, gpt], axis=0)

data

data['origin'] = 'AI'
data['model'] = 'chat gpt'
data = data.rename(columns = {'Text':'text'})
data = data.drop(columns = ['Unnamed: 0', 'Synthetic'])
data.head()

data.duplicated().sum()
data = data.drop_duplicates()

data.isna().sum()

df = pd.concat([df, data], axis=0)
df.info()

df['origin'].value_counts()

#deixando com 300.000 dados de twitter human
df_gpt = df[df["origin"] == "AI"]
df_amostra = df_gpt.sample(n=65393, random_state=42) 

df_amostra

# Remove as linhas do df que são comuns entre df e df_amostra
df = df.merge(df_amostra, how='left', indicator=True).query('_merge == "left_only"').drop(columns=['_merge'])
df

df['origin'].value_counts()

df.to_csv('df_pronto/df_gpt.csv', index=False)
