==> 1_human_df.ipynb <==
import pandas as pd

dataset = pd.read_json('database/tweetNERD.json', lines=True) #https://huggingface.co/datasets/socialmediaie/TweetNERD
dataset.head()

dataset = dataset.drop(columns = ['tweet_id', 'phrase', 'start', 'end', 'entityId', 'score', 'part'])
dataset = dataset.rename(columns = {'full_text':'text'})
dataset['origin'] = 'human'
dataset['model'] = None
dataset.head()

lista = ['tweet_eval_irony', 'tweet_eval_emotion'] # https://huggingface.co/datasets/cardiffnlp/tweet_eval/blob/main/README.md
for i in lista:
    x = pd.read_csv(f'database/{i}.csv')
    x = x.drop(columns = ['label'])
    x['origin'] = 'human'
    x['model'] = None
    dataset = pd.concat([x, dataset], axis=0)

lista = ['tweet_eval_abortion', 'tweet_eval_offencive','tweet_eval_hate', 'tweet_eval_sentiment', ] # https://huggingface.co/datasets/cardiffnlp/tweet_eval/blob/main/README.md
for i in lista:
    x = pd.read_parquet(f'database/{i}.parquet')
    x = x.drop(columns = ['label'])
    x['origin'] = 'human'
    x['model'] = None
    dataset = pd.concat([x, dataset], axis=0)

data = pd.read_csv('database/twitter_human.csv')
data['origin'] = 'human'
data['model'] = None
data = data.drop(columns = ['label', 'tweet_hashtags', 'datetime'])
data.head()

dataset = pd.concat([data, dataset], axis=0)
dataset.info()

data = pd.read_csv('database/Twitter_Data.csv') # https://huggingface.co/datasets/Balasuriya/Twitter_Data
data = data.rename(columns = {'clean_text':'text'})
data['origin'] = 'human'
data['model'] = None
data = data.drop(columns = ['category'])
data.head()

dataset = pd.concat([data, dataset], axis=0)
dataset.info()

dataset.isna().sum()
dataset.duplicated().sum()

dataset = dataset.dropna(subset=['text'])
dataset = dataset.drop_duplicates()

dataset.info()

data = pd.read_csv('database/sentiment140.csv', encoding='ISO-8859-1')
data = data.rename(columns = {"@switchfoot http://twitpic.com/2y1zl - Awww, that's a bummer.  You shoulda got David Carr of Third Day to do it. ;D" :"text"})
data['origin'] = 'human'
data['model'] = None
data = data.drop(columns = ['0', '1467810369', 'Mon Apr 06 22:19:45 PDT 2009', 'NO_QUERY','_TheSpecialOne_'])
data.head()

data.isna().sum()
data.duplicated().sum()

data = data.dropna(subset=['text'])
data = data.drop_duplicates()

#separando 44955 para completar os 500.000
df_amostra = data.sample(n=44955, random_state=42) 

dataset = pd.concat([df_amostra, dataset], axis=0)
dataset.info()

dataset.to_csv('df_pronto/df_human.csv', index=False)
