==> 2_test_feature_analytics.ipynb <==
import pandas as pd
import sys
import numpy as np
sys.path.append('/home/tammy.kojima/Authorship-attribution/')
import feature_extraction as fe
from tqdm import tqdm
import pickle
import os
import ast
from sklearn.model_selection import train_test_split

df = pd.read_csv('/home/tammy.kojima/Authorship-attribution/df_pronto/df_gpt_com_features.csv')
df.head()

df = df.drop(columns=["ai_perplexity_score", "ai_repeated_ngrams", "llm_chatgpt_patterns", "llm_structured_output", "llm_overly_polite", "llm_disclaimer_density"])
df.head()

df_gpt = pd.read_csv('df_pronto/df_gpt.csv')
df_human = pd.read_csv('df_pronto/df_human.csv')
#df_gpt = df_gpt.iloc[:10000]
df_human = df_human.iloc[:100000]
df_test = pd.concat([df_gpt, df_human], ignore_index=True)
df_test

# 2. Selecionar a coluna de textos
# (substitua 'texto' pelo nome real da coluna com o conteúdo)
textos = df_test["text"].fillna("").tolist()

# Tokeniza todos os textos primeiro
processed_texts = [fe.preprocess_text(texto) for texto in tqdm(textos, desc="Tokenizando textos")]

df_prossessed_texts = pd.DataFrame({'text': processed_texts})
df_prossessed_texts.to_csv('df_pronto/df_tokens_gpt.csv', index=False)

# Lista para armazenar resultados
lexical_features = []

for texto in tqdm(textos):
    # Tokeniza o texto usando Twokenizer
    words = fe.preprocess_text(texto)  # já retorna tokens em lowercase
    
    # Extrai features léxicas
    features = fe.extract_lexical_features(texto, words)
    lexical_features.append(features)

df_lexical = pd.DataFrame(lexical_features)

df_test = pd.concat([df_test, df_lexical], axis=1)
df_test.head()

stylistic_features = [
    fe.extract_stylistic_features(textos[i], len(processed_texts[i]))
    for i in tqdm(range(len(textos)), desc="Extraindo features estilísticas")
]
df_syntactic = pd.DataFrame(stylistic_features)
df_syntactic.head()

df_test = pd.concat([df_test, df_syntactic], axis=1)
df_test.head()

# especifico para casos com letra maiuscula 
df_test = pd.read_csv('df_pronto/df_2pronto.csv')
textos = df_test["text"].fillna("").tolist()
processed_texts = [fe.preprocess_text_A(texto) for texto in tqdm(textos, desc="Tokenizando textos")]
stylistic_features = [
    fe.extract_stylistic_featuresA(textos[i], len(processed_texts[i]))
    for i in tqdm(range(len(textos)), desc="Extraindo features estilísticas")
]
df_syntactic = pd.DataFrame(stylistic_features)
df_syntactic.head()

df_test = pd.concat([df_test, df_syntactic], axis=1)

structural_features = [
    fe.extract_structural_features(textos[i], processed_texts[i], len(processed_texts[i]))
    for i in tqdm(range(len(textos)), desc="Extraindo features estruturais")
]
df_structural = pd.DataFrame(structural_features)
df_structural.head()

df_test = pd.concat([df_test, df_structural], axis=1)
df_test.to_csv('df_pronto/df_3pronto.csv', index=False)

df_processed_texts = pd.read_csv('/home/tammy.kojima/Authorship-attribution/df_pronto/df_tokens_gpt.csv')
df_test = pd.read_csv('/home/tammy.kojima/Authorship-attribution/df_pronto/df_3pronto.csv')
processed_texts = df_processed_texts['processed_text'].apply(ast.literal_eval).tolist()
textos = df_test["text"].fillna("").tolist()

# falta fazer esse 
syntactic_features = [
    fe.extract_syntactic_features(tokens)
    for tokens in tqdm(processed_texts, desc="Extraindo features sintáticas")
]
df_syntactic = pd.DataFrame(syntactic_features)
df_syntactic.head()

df_test = pd.concat([df_test, df_syntactic], axis=1)
df_test.head()

df_processed_texts = pd.read_csv('/home/tammy.kojima/Authorship-attribution/df_pronto/df_tokens_gpt.csv')

processed_texts = df_processed_texts['processed_text'].apply(ast.literal_eval).tolist()
textos = df["text"].fillna("").tolist()

X_new_features = fe.extract_features_batch(textos, processed_texts = processed_texts)

df_features = pd.DataFrame(X_new_features)
df_features.head()

df.head()

df_test = pd.concat([df, df_features], axis=1)
df_test.head()

# Exemplo de uso básico
ngrams_df = fe.extract_ngrams_features(processed_texts, max_features=300)
ngrams_df.head()

# Resetar os índices de ambos os DataFrames
df_test_reset = df.reset_index(drop=True)
ngrams_df_reset = ngrams_df.reset_index(drop=True)

# Agora concatenar
df_combined = pd.concat([df_test_reset, ngrams_df_reset], axis=1)

df_test_reset = df.reset_index(drop=True)

# Troca IA → 1 e human → 0
df_test_reset['origin'] = df_test_reset['origin'].replace({'AI': 1, 'human': 0})

df_test_reset.head()

df_combined = df_combined.apply(lambda x: x.sparse.to_dense() if pd.api.types.is_sparse(x) else x)
df_combined.describe()

df_combined.isna().sum()

df_test_reset.to_csv("/home/tammy.kojima/Authorship-attribution/df_pronto/df_gpt_com_features.csv", index=False)
